move provider into the same file of its hook (since it is not JSX anymore)
