+++ src/pages/namespace/Instances/Detail/index.tsx
@@ -4,21 +4,20 @@ import {
   useInstanceId,
 } from "./store/instanceContext";
 
-import { InstanceStreamingProvider } from "~/api/instances/query/InstanceStreamingProvider";
+import { InstanceStreamingProvider } from "~/api/instances/query/details";
 import InstancesDetail from "./InstanceDetail";
-import { LogStreamingProvider } from "~/api/logs/query/LogStreamingProvider";
+import { LogStreamingProvider } from "~/api/logs/query/get";
 import { pages } from "~/util/router/pages";
 
 const InstanceStreaming = () => {
   const instanceId = useInstanceId();
   const filters = useFilters();
-
   return (
-    <InstanceStreamingProvider instanceId={instanceId}>
-      <LogStreamingProvider filters={filters} instanceId={instanceId}>
-        <InstancesDetail />
-      </LogStreamingProvider>
-    </InstanceStreamingProvider>
+    <>
+      <InstanceStreamingProvider instanceId={instanceId} />
+      <LogStreamingProvider filters={filters} instanceId={instanceId} />
+      <InstancesDetail />
+    </>
   );
 };
